[CodeGen][AMDGPU] Move boilerplate unittest code to base clas (NFC)
This adds the `CodeGenTestBase` class to handle boilerplate code for
codegen unitests and makes use of it wherever possible, in particular in
AMDGPU unit tests.

Furthermore, this makes all AMDGPU unit tests rely on GoogleTest's
API for "run once per testsuite" code, instead of re-implenting that
behavior using a `std::once` flag. As a consequence all TEST(...) become
TEST_F().
